refactor: Update backend imports and class names for consistency
