+++ src/qubex/backend/__init__.py
@@ -1,9 +1,11 @@
 from .config_loader import ConfigLoader
 from .control_system import Box, BoxType, ControlSystem, Port
-from .experiment_system import ControlParams, ExperimentSystem, Target
+from .experiment_system import ControlParams, ExperimentSystem
 from .lattice_graph import LatticeGraph
 from .quantum_system import Chip, QuantumSystem, Qubit, Resonator
 from .qube_backend import SAMPLING_PERIOD, QubeBackend, QubeBackendResult
+from .qubecalib_manager import QubeCalibManager
+from .target import Target, TargetType
 
 __all__ = [
     "Box",
@@ -18,8 +20,10 @@ __all__ = [
     "QubeBackendResult",
     "ControlSystem",
     "QuantumSystem",
+    "QubeCalibManager",
     "Qubit",
     "Resonator",
     "SAMPLING_PERIOD",
     "Target",
+    "TargetType",
 ]
